Fix message text overflow in statusMessage
The text content inside the "statusMessage" overflows and crosses the bounding-box of "statusMessage" div.  

So, I have added style args to statusMessage div element. This ensures that the textContent inside the statusMessage div remains within the bounds of it's parent div. 

Now, it works fine.



Alternatively we can also create a css-rule in appropriate stylesheet also.
#statusMessage
{
   white-space: pre-wrap;      /* CSS3 */   
   white-space: -moz-pre-wrap; /* Firefox */    
   white-space: -pre-wrap;     /* Opera <7 */   
   white-space: -o-pre-wrap;   /* Opera 7 */    
   word-wrap: break-word;      /* IE */
}